Break root routes into separate file.
